+++ OOP/OOP_PrioritizeMyFuture_2023/ilya.json
@@ -1,1 +1,1 @@
-[["hello", 3, "gfd", "Not Done"], ["vfd", 5, "bgf", "Not Done"]]
+[["hello", 3, "vfd", "Not Done"]]
