+++ OOP/OOP_PrioritizeMyFuture_2023/requirements.txt
@@ -1,0 +1,2 @@
+passlib~=1.7.4
+setuptools~=65.5.1
