+++ OOP/OOP_PrioritizeMyFuture_2023/setup.py
@@ -1,0 +1,10 @@
+from setuptools import find_packages, setup
+setup(
+    name='mypythonlib',
+    packages=find_packages(include=['mypythonlib']),
+    version='0.1.0',
+    description='My first Python library(DATA/SE/DE/RIALIZATION)',
+    author='Me',
+    license='MIT',
+    install_requires=[],
+)
